Experiments with Logistic Regression
Used Logistic Regression with use of PyTorch as machine learning method. Results for tournaments are as always written in sheet in ./aiStatistics. The training is about 6 seconds long, so it can manage the 10 seconds limit for constructing of AI before each game. I think this machine learning method could be good for submission of our project.